+++ lib/jade/stdlib.rb
@@ -107,7 +107,14 @@ module Jade
     def add_imports(entry)
       (STDLIBS - EXTENSIONS)
         .reduce(entry) do |acc, stdlib|
-          ImportEntry[stdlib.entry.name, stdlib.entry.name, stdlib.default_imports, stdlib.entry.exposes]
+          ImportEntry
+            .new(
+              module_name: stdlib.entry.name,
+              alias: stdlib.entry.name,
+              unqualified_symbols: stdlib.default_imports,
+              qualified_symbols: stdlib.entry.exposes,
+              implicit: true,
+            )
             .then { acc.import(it) }
         end
     end