+++ lib/jade/stdlib/intrinsics.rb
@@ -285,7 +285,14 @@ module Jade
         # TODO: This is the same code from stdlib that auto imports stuff.
         imports
           .reduce(entry) do |acc, stdlib|
-            ImportEntry[stdlib.entry.name, stdlib.entry.name, stdlib.default_imports, stdlib.entry.exposes]
+            ImportEntry
+              .new(
+                module_name: stdlib.entry.name,
+                alias: stdlib.entry.name,
+                unqualified_symbols: stdlib.default_imports,
+                qualified_symbols: stdlib.entry.exposes,
+                implicit: true,
+              )
               .then { acc.import(it) }
           end
       end