+++ lib/jade/symbol/ambiguous.rb
@@ -1,0 +1,5 @@
+module Jade
+  module Symbol
+    Ambiguous = Data.define(:name, :modules)
+  end
+end