+++ lib/jade/frontend/type_checking/inference/helpers.rb
@@ -36,6 +36,17 @@ module Jade
 
             result
           end
+
+          def type_mismatch(state, node)
+            ->(error) do
+              Error::TypeMismatch.new(
+                state.env.entry_name,
+                node.range,
+                expected: error.expected,
+                actual: error.actual,
+              )
+            end
+          end
         end
       end
     end